업데이트
업데이트

+++ Project/Assets/Script/ForSkill/Skill.cs
@@ -6,7 +6,7 @@ public class Skill : MonoBehaviour
 {
 
     public string Skillname;
-    public int SkillDamage;
+    public float SkillDamage;
     public int SkillRange;
     public bool CanUseSkill ;
     public float SkillCoolTime;
